starting to add docs
